Allow rewrite elaboration
Summary:
The left-hand side of a rewrite is an expression/type/pattern. We have a tool
for rewriting expressions/types/patterns. Lets use rewrites to rewrite
rewrites!

Seriously though, this is useful. The prototypical example is elaborating the
dollar operator to parentheses. If the user specifies the following rewrite:

    forall x. print $ x = putStrLn $ show x

Then it will only match explicit uses of the dollar operator. The expression
`print (foo bar)` will NOT get rewritten, even though it is equivalent to
`print $ foo bar`.

We had a hack in the PatternMap instance for
expressions that effectively applied the rewrite `forall f x. f $ x = f (x)` to
the left-hand side of all expression rewrites. Now there is a nice way to define these kinds of
elaborations in general.

Adds four command-line flags (`--elaborate`, `--elaborate-type`,
`--elaborate-pattern`, and `--no-default-elaborations`) and a set of
elaborations to use by default.

Currently the default set is just the dollar rewrite. Would be nice to add
some others.

Reviewed By: pepeiborra

Differential Revision: D28107722

fbshipit-source-id: 141e25e508f4fc1df7e1f94009df43bf72206160