Remove dependency on xargs (#31)
Summary:
Windows support currently is blocked on a series of escaping issues caused by xargs, as discussed in https://github.com/facebookincubator/retrie/issues/21 . This pull request replaces the dependency on xargs by a series of separate grep calls.

The good news is that this does allow usage from haskell-language-server on Windows. The bad news is that direct calls to the retrie executable still break non deterministically. Sometimes retrie works correctly, frequently it errors with errors like

    C:\sr\snapshots\6bf6c8f4\pkgdb\package.cache.lock: openBinaryFile: resource busy (file is locked)

Occasionally the output gets mangled badly as well, though the underlying error message appears to be the same

    C:\\Users\Name\ACp:p\Dsart\as\nLaopcsahlo\tPsr\o6gbrfa6mcs8\fs4t\apckkg\dxb8\6p_a6c4k-awgien.dcoawcsh\eg.hlco-c8k.:1 0o.p4e\nlBiibn\apraycFkialgee:. croensfo.udr\cpea cbkuasgy e .(cfaiclhee .ilso clko:c koepde)n

I'm reasonably sure that these errors aren't my fault but I'd be thankful if someone with access to a linux system could check that the test suite didn't regress on linux.

Pull Request resolved: https://github.com/facebookincubator/retrie/pull/31

Reviewed By: simonmar

Differential Revision: D30156688

Pulled By: pepeiborra

fbshipit-source-id: 665b6e97937a10526438bc00a189710f45c0948b