escape single quote in pattern when calling grep
It's not sufficient to escape characters that are special to grep.
One must also escape the single quote that is used to quote the
pattern to the shell.  I've used the classic POSIX trick of
replacing the single ' with '"'"'.

To expose the issue in the original code, try something on this order:

    retrie --dry-run --adhoc "forall x . backendNeedn'tLink x = not (backendNeedsLink x)"