Set `buildable: false` for unsupported ghc
Currently the retrieval `1.2.1` in hackage shadows versions that work with older compilers.  The cabal planner selects this version (`1.2.1`) because it does not think it needs the `ghc` library at all when making a plan for older versions of ghc.

I think we may also need to deprecate `1.2.1` in hackage (as I don't know if this change can be made as a hackage revision).  Perhaps another option might be:

```
  build-depends: ghc >= 9.2
  if impl (ghc >= 9.4) && impl (ghc < 9.5)
    build-depends:
       ghc == 9.4.*,
       ghc-exactprint >= 1.6.0 && < 1.7
  else
    if impl (ghc >= 9.2) && impl (ghc < 9.3)
      build-depends:
        ghc == 9.2.*,
        ghc-exactprint < 1.6.0 && > 1.4.0
```