Explicitly store the original pattern location in Query
This is used to filter out potential matches in the file the query
originates in.

But often we have had to re-arrange it with deltas at the point of
matching, so can no longer easily extract it from the pattern to be
matched.

Instead we store it explicitly when making a Query